add discord role check and give logic
