+++ explorer/src/utils/auth/vcs/discord/role-nominator.ts
@@ -18,15 +18,6 @@ export const verifyTaurusDiscordNominatorRole = async (roles: string[]) => {
   return roles.includes(TAURUS_DISCORD_GUILD_ROLE_ID_NOMINATOR)
 }
 
-export const verifyMainnetDiscordTalismanFarmerRole = async (roles: string[]) => {
-  if (!process.env.MAINNET_DISCORD_GUILD_ROLE_ID_TAILSMAN_FARMER)
-    throw new Error('No Discord guild role ID for talisman farmer')
-  const { MAINNET_DISCORD_GUILD_ROLE_ID_TAILSMAN_FARMER } = process.env
-
-  // Check if the user has the nominator role
-  return roles.includes(MAINNET_DISCORD_GUILD_ROLE_ID_TAILSMAN_FARMER)
-}
-
 export const giveMainnetDiscordNominatorRole = async (userId: string) => {
   if (!process.env.MAINNET_DISCORD_GUILD_ROLE_ID_NOMINATOR)
     throw new Error('No Discord guild role ID for nominator')
@@ -50,18 +41,5 @@ export const giveTaurusDiscordNominatorRole = async (userId: string) => {
     userId,
     TAURUS_DISCORD_GUILD_ROLE_ID_NOMINATOR,
     'Give the user the nominator role',
-  )
-}
-
-export const giveMainnetDiscordTalismanFarmerRole = async (userId: string) => {
-  if (!process.env.MAINNET_DISCORD_GUILD_ROLE_ID_TAILSMAN_FARMER)
-    throw new Error('No Discord guild role ID for talisman farmer')
-  const { MAINNET_DISCORD_GUILD_ROLE_ID_TAILSMAN_FARMER } = process.env
-
-  // Add the talisman farmer role to the user
-  await giveDiscordRole(
-    userId,
-    MAINNET_DISCORD_GUILD_ROLE_ID_TAILSMAN_FARMER,
-    'Give the user the talisman farmer role',
   )
 }
